requirements.txt
acrcloud==1.1
aiohttp==3.5.4
appnope==0.1.0
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
backcall==0.1.0
beautifulsoup4==4.8.0
bleach==3.1.0
bs4==0.0.1
certifi==2018.11.29
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
decorator==4.4.0
defusedxml==0.5.0
Django==2.1.7
entrypoints==0.3
feedparser==5.2.1
future==0.17.1
gitdb2==2.0.5
GitPython==2.1.13
idna==2.8
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
iso8601==0.1.12
jedi==0.13.3
Jinja2==2.10.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
libmagic==1.0
markdownify==0.4.1
MarkupSafe==1.1.1
mistune==0.8.4
multidict==4.5.2
nbconvert==5.4.1
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.2
pandocfilters==1.4.2
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pyacrcloud==1.0.1
PyAutoGUI==0.9.44
pycparser==2.19
PyGetWindow==0.0.6
Pygments==2.3.1
PyMouse==1.0
PyMsgBox==1.0.7
PyRect==0.1.4
pyrsistent==0.14.11
PyScreeze==0.1.21
pyserial==3.4
python-dateutil==2.8.0
python-telegram-bot==12.0.0
PyTweening==1.0.3
pytz==2018.9
PyYAML==5.1
pyzmq==18.0.1
qtconsole==4.4.3
requests==2.21.0
Send2Trash==1.5.0
six==1.12.0
smmap2==2.0.5
soupsieve==1.9.3
telepot==12.7
terminado==0.8.2
testpath==0.4.2
tornado==6.0.2
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
yarl==1.3.0
